Post-rebase fix-ups
